test: update flutter unit tests.

+++ webf/test/src/css/display_flex_test.dart
@@ -227,8 +227,8 @@ void main() {
       final flex = prepared.getElementById('flex');
       final bigCjk = prepared.getElementById('big-cjk');
 
-      final int offsetWidth = flex.offsetWidth;
-      final int scrollWidth = flex.scrollWidth;
+      final double offsetWidth = flex.offsetWidth;
+      final double scrollWidth = flex.scrollWidth;
 
       expect(scrollWidth, lessThanOrEqualTo(offsetWidth + 1), reason: 'flex container should not overflow horizontally');
       expect(bigCjk.offsetWidth, lessThan(100), reason: 'CJK item should shrink when flexing within constraints');
@@ -594,10 +594,10 @@ void main() {
 
       // Container should respect min-height
       expect(container.offsetHeight, equals(100.0));
-      
+
       // Red div should stretch to container height (98px), not green div height (50px)
       expect(redDiv.offsetHeight, equals(98.0));
-      
+
       // Green div should keep its explicit height
       expect(greenDiv.offsetHeight, equals(50.0));
     });
@@ -639,12 +639,12 @@ void main() {
 
       // Container should be at least min-width (200px), may be larger due to viewport
       expect(container.offsetWidth, greaterThanOrEqualTo(200.0));
-      
+
       // Red div should stretch to container content width (container width - 2px border), not green div width (80px)
       double expectedRedWidth = container.offsetWidth - 2.0; // Account for border
       expect(redDiv.offsetWidth, equals(expectedRedWidth));
       expect(redDiv.offsetWidth, greaterThan(80.0)); // Should be wider than green div
-      
+
       // Green div should keep its explicit width
       expect(greenDiv.offsetWidth, equals(80.0));
     });
@@ -694,7 +694,7 @@ void main() {
       // 2 × 5px padding = 10px padding
       // 2 × 1px border = 2px border
       // Total = 122px
-      
+
       expect(container.offsetWidth, equals(122.0));
     });
 
@@ -738,10 +738,10 @@ void main() {
 
       // Expected calculation without gap:
       // 3 children × 30px = 90px content
-      // 2 × 5px padding = 10px padding  
+      // 2 × 5px padding = 10px padding
       // 2 × 1px border = 2px border
       // Total = 102px
-      
+
       expect(container.offsetWidth, equals(102.0));
     });
 
@@ -752,7 +752,7 @@ void main() {
         html: '''
           <html>
             <body style="margin: 0; padding: 0; font-size: 16px;">
-              Items with gap: 
+              Items with gap:
               <div id="container" style="
                 display: inline-flex;
                 gap: 10px;
@@ -794,17 +794,17 @@ void main() {
       final redRect = redDiv.getBoundingClientRect();
       final greenRect = greenDiv.getBoundingClientRect();
       final blueRect = blueDiv.getBoundingClientRect();
-      
+
       print('=== DEBUG INLINE-FLEX OVERFLOW ===');
       print('Container offsetWidth: ${container.offsetWidth}');
       print('Container rect: left=${containerRect.left}, right=${containerRect.right}, width=${containerRect.width}');
       print('Red rect: left=${redRect.left}, right=${redRect.right}');
-      print('Green rect: left=${greenRect.left}, right=${greenRect.right}'); 
+      print('Green rect: left=${greenRect.left}, right=${greenRect.right}');
       print('Blue rect: left=${blueRect.left}, right=${blueRect.right}');
       print('Blue overflows container? ${blueRect.right > containerRect.right}');
-      
+
       // The key test: blue item should NOT overflow the container
-      expect(blueRect.right, lessThanOrEqualTo(containerRect.right), 
+      expect(blueRect.right, lessThanOrEqualTo(containerRect.right),
              reason: 'Blue item should not overflow the lightyellow container');
     });
   });
